Refactor the Cucumber hooks into module functions to be a bit easier to test.
In order to test these things effectively, we need to pull in some magic code
straight from the cucumber code-base itself.

As far as I can tell, this is the only "reasonable" way to test hooks or things
of that ilk because of the tendency for "cucumber magic" to happen inside of a
weird global 'World' scope

I could just be misunderstanding all of this though.