Add RSpec parallelization
Added a Rake task for running tests in parallel, 1 thread per concurrency available on SL.

TestBroker handles the number of tests, the platform brokerage and the concurrency.

Relying on the "worker" facility of parallel_tests to dispatch tests to available Sauce VMs.

Saucerspec runner replaces the rspec runner for parallel_tests.  The runner creates a duplicate test group for each set of environment variables and also attaches variables from the broker for each group.

Config now only returns the first entry of browsers if not running in || tests.

Increased version to 3