Use TEST_ENV_NUMBER to set Rails port Properly return TEST_ENV_NUMBER
Test_env_number is based on the current process number, which is based on the index of the group in the group list... Which isn't going to work so well for multiple identical groups.

Extended the broker to provide correct group numbering, made the Rails server start on various ports based off this.