Try to guess if Capybara is in use when using #page.
If the Capybara::DSL module is included, *always* return #page.

Note the tests don't pass with current versions of Capybara because when
sessions are reset, we close off the browser but Capybara then makes
assertions on it.

In current versions of Capybara, the way drivers are reset also means we
try to re-use drivers in new tests.  Versions of Capybara > 2.2.1 no
longer have this problem (At least, once it's released >.>)