Add hooks for post-job actions.
Added hooks for post job actions.  Hooks are registered in your config
block against a keyword using the `after_job` method, which takes a
symbol and a block requiring 3 values, an id, a name and whether or not
the job was successful.

There is a default action taken for :rspec and :jenkins (where present).
Registered an action against these hooks will over-write the original
action.

```ruby
Sauce.config do |c|
  c.after_job :rspec do |id, name, success|
    STDERR.puts "Job #{id} succeeded? -- #{success}
  end
end