Nullifying Sauce.driver_pool
Need to nullify this as well otherwise rspec_browser returns an object which can't be used because it's pointing at the  closed session. This happens when you run background tasks in cucumber. Previous scenario closes the browser, and in your background you try to interact with a page, it tries to use old browser and thorws exception.