Sauce Connect - Quiet and Timeouty.
Allowed for the @quiet parameter to be passed to Sauce Connect.

Allowed for the timeout before Sauce Connect freaks out to be set as
part of the Sauce Config.

Changed the code when waiting for a tunnel started by ||tests.  Tests
now wait for `Sauce::Utilities::Connect::TIMEOUT` and will raise
`Sauce::Utilities::Connect::TunnelNeverStarted` exceptions unless, uh,
the tunnel starts.